examples and unit tests run with the reorganization of mixins
