+++ examples/galSimGalaxyBulgeGenerator.py
@@ -4,7 +4,8 @@ galaxy bulges (or any sersic profile)
 """
 
 import os
-from lsst.sims.catalogs.generation.db import CatalogDBObject, ObservationMetaData
+from lsst.sims.catalogs.generation.db import CatalogDBObject
+from lsst.sims.utils import ObservationMetaData
 from lsst.sims.catUtils.baseCatalogModels import GalaxyBulgeObj, OpSim3_61DBObject
 from lsst.sims.GalSimInterface import GalSimGalaxies, SNRdocumentPSF
